LoggerProject/requirements.txt
astroid==2.1.0
gitdb2==2.0.5
GitPython==2.1.11
isort==4.3.4
lazy-object-proxy==1.3.1
LoggerProject==0.0.0.1+git
mccabe==0.6.1
peewee==3.8.1
pylint==2.2.2
pymongo==3.7.2
pytz==2018.7
six==1.12.0
smmap2==2.0.5
typed-ast==1.2.0
wrapt==1.11.1
